+++ pyproject.toml
@@ -29,17 +29,27 @@ dependencies = [
 dynamic = ["version", "description", "authors", "urls", "keywords"]
 
 [project.optional-dependencies]
-lab = ["jupyterlab>=4.2.0", "jupyter-docprovider>=1.0.0b1", "jupyter-server-ydoc>=1.0.0b1"]
+lab = [
+  "jupyterlab>=4.2.0",
+  "jupyter-docprovider>=1.0.0b1",
+  "jupyter-server-ydoc>=1.0.0b1",
+]
 test = [
   "pytest~=8.2",
   "pytest-cov",
   "pytest-jupyter[server]>=0.6",
   "pytest-timeout",
   "jupyter-server-ydoc[test]>=1.0.0b1",
-  "jupyter-server-fileid"
+  "jupyter-server-fileid",
 ]
-lint = ["mdformat>0.7", "mdformat-gfm>=0.3.5", "ruff>=0.4.0"]
-typing = ["mypy>=0.990"]
+lint = [
+  "mdformat>0.7",
+  "mdformat-gfm>=0.3.5",
+  "ruff>=0.4.0",
+]
+typing = [
+  "mypy>=0.990",
+]
 
 [tool.hatch.version]
 source = "nodejs"
