+++ src/executor.ts
@@ -4,29 +4,14 @@
  * Distributed under the terms of the Modified BSD License.
  */
 
-import {
-  JupyterFrontEnd,
-  JupyterFrontEndPlugin
-} from '@jupyterlab/application';
 import { Dialog, showDialog } from '@jupyterlab/apputils';
-import {
-  CodeCell,
-  type ICodeCellModel,
-  type MarkdownCell
-} from '@jupyterlab/cells';
+import { CodeCell } from '@jupyterlab/cells';
+import type { ICodeCellModel, MarkdownCell } from '@jupyterlab/cells';
 import { URLExt } from '@jupyterlab/coreutils';
 import { INotebookCellExecutor } from '@jupyterlab/notebook';
-import { OutputPrompt, Stdin } from '@jupyterlab/outputarea';
-import { Kernel, ServerConnection } from '@jupyterlab/services';
-import * as KernelMessage from '@jupyterlab/services/lib/kernel/messages';
-import { nullTranslator, type ITranslator } from '@jupyterlab/translation';
-import { PromiseDelegate } from '@lumino/coreutils';
-import { Panel } from '@lumino/widgets';
-
-/**
- * Polling interval for accepted execution requests.
- */
-const MAX_POLLING_INTERVAL = 1000;
+import { ServerConnection } from '@jupyterlab/services';
+import { nullTranslator } from '@jupyterlab/translation';
+import { requestServer } from './requestServer';
 
 /**
  * Notebook cell executor posting a request to the server for execution.
@@ -107,7 +92,7 @@ export class NotebookCellServerExecutor implements INotebookCellExecutor {
           }
 
           const kernelId = sessionContext?.session?.kernel?.id;
-          const apiURL = URLExt.join(
+          const executeApiURL = URLExt.join(
             this._serverSettings.baseUrl,
             `api/kernels/${kernelId}/execute`
           );
@@ -133,7 +118,7 @@ export class NotebookCellServerExecutor implements INotebookCellExecutor {
             // FIXME quid of deletedCells and timing record
             const response = await requestServer(
               cell as CodeCell,
-              apiURL,
+              executeApiURL,
               init,
               this._serverSettings,
               translator
@@ -167,175 +152,4 @@ export class NotebookCellServerExecutor implements INotebookCellExecutor {
   }
 }
 
-async function requestServer(
-  cell: CodeCell,
-  url: string,
-  init: RequestInit,
-  settings: ServerConnection.ISettings,
-  translator?: ITranslator,
-  interval = 100
-): Promise<Response> {
-  const promise = new PromiseDelegate<Response>();
-  ServerConnection.makeRequest(url, init, settings)
-    .then(async response => {
-      if (!response.ok) {
-        if (response.status === 300) {
-          let replyUrl = response.headers.get('Location') || '';
-
-          if (!replyUrl.startsWith(settings.baseUrl)) {
-            replyUrl = URLExt.join(settings.baseUrl, replyUrl);
-          }
-          const { parent_header, input_request } = await response.json();
-          // TODO only the client sending the snippet will be prompted for the input
-          // we can have a deadlock if its connection is lost.
-          const panel = new Panel();
-          panel.addClass('jp-OutputArea-child');
-          panel.addClass('jp-OutputArea-stdin-item');
-
-          const prompt = new OutputPrompt();
-          prompt.addClass('jp-OutputArea-prompt');
-          panel.addWidget(prompt);
-
-          const input = new Stdin({
-            future: Object.freeze({
-              sendInputReply: (
-                content: KernelMessage.IInputReply,
-                parent_header: KernelMessage.IHeader<'input_request'>
-              ) => {
-                ServerConnection.makeRequest(
-                  replyUrl,
-                  {
-                    method: 'POST',
-                    body: JSON.stringify({ input: content.value })
-                  },
-                  settings
-                ).catch(error => {
-                  console.error(
-                    `Failed to set input to ${JSON.stringify(content)}.`,
-                    error
-                  );
-                });
-              }
-            }) as Kernel.IShellFuture,
-            parent_header,
-            password: input_request.password,
-            prompt: input_request.prompt,
-            translator
-          });
-          input.addClass('jp-OutputArea-output');
-          panel.addWidget(input);
-
-          // Get the input node to ensure focus after updating the model upon user reply.
-          const inputNode = input.node.getElementsByTagName('input')[0];
-
-          void input.value.then(value => {
-            panel.addClass('jp-OutputArea-stdin-hiding');
-
-            // FIXME this is not great as the model should not be modified on the client.
-            // Use stdin as the stream so it does not get combined with stdout.
-            // Note: because it modifies DOM it may (will) shift focus away from the input node.
-            cell.outputArea.model.add({
-              output_type: 'stream',
-              name: 'stdin',
-              text: value + '\n'
-            });
-            // Refocus the input node after it lost focus due to update of the model.
-            inputNode.focus();
-
-            // Keep the input in view for a little while; this (along refocusing)
-            // ensures that we can avoid the cell editor stealing the focus, and
-            // leading to user inadvertently modifying editor content when executing
-            // consecutive commands in short succession.
-            window.setTimeout(async () => {
-              // Tack currently focused element to ensure that it remains on it
-              // after disposal of the panel with the old input
-              // (which modifies DOM and can lead to focus jump).
-              const focusedElement = document.activeElement;
-              // Dispose the old panel with no longer needed input box.
-              panel.dispose();
-              // Refocus the element that was focused before.
-              if (focusedElement && focusedElement instanceof HTMLElement) {
-                focusedElement.focus();
-              }
-
-              try {
-                const response = await requestServer(
-                  cell,
-                  url,
-                  init,
-                  settings,
-                  translator
-                );
-                promise.resolve(response);
-              } catch (error) {
-                promise.reject(error);
-              }
-            }, 500);
-          });
-
-          cell.outputArea.layout.addWidget(panel);
-        } else {
-          promise.reject(await ServerConnection.ResponseError.create(response));
-        }
-      } else if (response.status === 202) {
-        let redirectUrl = response.headers.get('Location') || url;
-
-        if (!redirectUrl.startsWith(settings.baseUrl)) {
-          redirectUrl = URLExt.join(settings.baseUrl, redirectUrl);
-        }
-
-        setTimeout(
-          async (
-            cell: CodeCell,
-            url: string,
-            init: RequestInit,
-            settings: ServerConnection.ISettings,
-            translator?: ITranslator,
-            interval?: number
-          ) => {
-            try {
-              const response = await requestServer(
-                cell,
-                url,
-                init,
-                settings,
-                translator,
-                interval
-              );
-              promise.resolve(response);
-            } catch (error) {
-              promise.reject(error);
-            }
-          },
-          interval,
-          cell,
-          redirectUrl,
-          { method: 'GET' },
-          settings,
-          translator,
-          // Evanescent interval
-          Math.min(MAX_POLLING_INTERVAL, interval * 2)
-        );
-      } else {
-        promise.resolve(response);
-      }
-    })
-    .catch(reason => {
-      promise.reject(new ServerConnection.NetworkError(reason));
-    });
-  return promise.promise;
-}
-
-export const notebookCellExecutor: JupyterFrontEndPlugin<INotebookCellExecutor> =
-  {
-    id: 'jupyter-server-nbmodel:notebook-cell-executor',
-    description:
-      'Add notebook cell executor that uses REST API instead of kernel protocol over WebSocket.',
-    autoStart: true,
-    provides: INotebookCellExecutor,
-    activate: (app: JupyterFrontEnd): INotebookCellExecutor => {
-      return new NotebookCellServerExecutor({
-        serverSettings: app.serviceManager.serverSettings
-      });
-    }
-  };
+export default NotebookCellServerExecutor;
